backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
gql[httpx]>=3.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pandas>=2.1.0
pyarrow>=14.0.0
